Using 'maintenance' mode to distinguish unkown nodes
This addresses the artifical scenario specific to VMs. Although we
have nodes registered in ironic, they aren't 'discovered' yet. Since
we need to make a blacklist of 'discovered' nodes, we set maintenance
mode to True on 'unkown' nodes. These are then skipped from the
blacklist and can then be served by discovery dnsmasq.